Move vnni check func into utils
