+++ torchao/kernel/intmm.py
@@ -10,7 +10,11 @@ import torch
 from torch._dynamo import is_compiling as dynamo_is_compiling
 from torch._higher_order_ops.out_dtype import out_dtype
 
-from torchao.utils import check_cpu_version, torch_version_at_least
+from torchao.utils import (
+    _cpu_is_vnni_supported,
+    check_cpu_version,
+    torch_version_at_least,
+)
 
 logger = logging.getLogger(__name__)
 logger.addHandler(logging.NullHandler())
@@ -115,19 +119,6 @@ def _cpu_is_amx_tile_supported() -> bool:
         return torch._C._cpu._is_amx_tile_supported()
     elif hasattr(torch.cpu, "_is_amx_tile_supported"):
         return torch.cpu._is_amx_tile_supported()
-    return False
-
-
-def _cpu_is_vnni_supported() -> bool:
-    """
-    Safely query AVX512_VNNI support, guarding against private API absence.
-    torch.cpu._is_vnni_supported / torch._C._cpu._is_vnni_supported are
-    private and may be missing in certain PyTorch builds or versions.
-    """
-    if hasattr(torch._C._cpu, "_is_vnni_supported"):
-        return torch._C._cpu._is_vnni_supported()
-    elif hasattr(torch.cpu, "_is_vnni_supported"):
-        return torch.cpu._is_vnni_supported()
     return False
